+++ scripts/macros/monsterFeatures/zombie/zombie.js
@@ -1,0 +1,4 @@
+import {undeadFortitude} from './undeadFortitude.js';
+export let zombie = {
+    'undeadFortitude': undeadFortitude
+}